cup_detection/messaging/requirements.txt
paho-mqtt==1.6.1
pyspark==3.5.0